Compose: Setup traefik to work as a proxy to allow for custom domains
Currently the frontend needs to be recompiled in order to host it on a
different domain since the "NEXT_PUBLIC_API_SERVER_URL" is compiled into
the dashboard.

If we instead compile the frontend to make requests to the backend on a
fixed root path (e.g. "/api") then we can use a proxy to forward those
requests to the backend without needing to worry about CORS.

This patch introduces traefik as a reverse proxy to give an idea on how
this would look like