getMetricsTags cached using a 600s redis cache (#92)
closes #91

Add a first draft of a cache for getMetricsTags - this uses simple inline architecture but could be refactored into a separate model if necessary. Could also use unit and/or functional tests (no tests in the API package right now ⁉️  )

Tested locally but definitely a candidate for further exploration. It also preserves the clickhouse / query metadata which is probably not correct.

Adds two new env configuration parameters:
    CACHE_METRICS_TAGS // boolean, disables metric cache if set to 'false'
    CACHE_METRICS_EXPIRATION // string, time description e.g. '1d', '1h', '600s' - duration of cache TTL

These could in theory be unified to take e.g. a cache expiration time or 'false'