improve saved search api + UI (#269)
Our saved search UI had a CSS regression and old branding, minor changes to fix it
<img width="853" alt="image" src="https://github.com/hyperdxio/hyperdx/assets/2781687/0509d165-242a-458c-a578-b31ee364e3af">

Additionally we had a bug in our PATCH endpoint not allowing empty query updates. I've also added an optional name parameter to make it easy to rename in the FE in the future. Moved things to zod validators as well.