Generic Webhooks v0 (#337)
[Generic Webhook Option for Alerts
](https://github.com/hyperdxio/hyperdx/issues/2#issue-1893833428
)

### DETAILS:

This PR enables the creation and use of generic webhooks alongside the existing slack webhooks. This allows users to configure arbitrary webhook consumers/destinations with custom payloads.

For now the lack of signage/security features means more complex webhooks that perform actions on alert will most likely be gated off due to their internal requirements, but this should unlock a variety of message-focused consumers alongside the existing slack implementation. Query parameter usage was built into the migration and logic, and can be enabled in a later version when security options make those more complex use cases (like caching) worthwhile. For the time being many consumers allow/mirror QP functionality in the body of the request, and otherwise building into the url manually achieves the same purpose.

This implementation assumes and is limited to POST requests only, which is the ideal sender behavior and has exceptionally large coverage, but optionality for GETs and PUTs can be added in later versions if they are desired.

Message templating is still quite limited while the more robust templating system is in development, and users should refer to their specific consumer documentation for implementation.

As a minor addition, with the added complexity beyond just single slack webhooks, optional descriptions were also added to the webhook model and displayed on the settings page.

### V1+ NEXT STEPS:
- security/signature functionality
- user facing webhook edit functionality
- functionality to send webhook tests during creation
- alignment with current in-progress alert templating
- user facing queryParam functionality (and/or url building for ease of use)

### VISUALS:

**TEAM SETTINGS UPDATE:**
![Screenshot 2024-03-07 at 3 16 11 PM](https://github.com/hyperdxio/hyperdx/assets/4743932/e0c3a60c-d9b6-4893-a8bb-81a4c6c829ea)

**GENERIC WEBHOOK CREATION:**
![Screenshot 2024-03-07 at 3 17 50 PM](https://github.com/hyperdxio/hyperdx/assets/4743932/ffc0945b-2dae-444b-9ef2-0e0d705c0c3d)

**ALERT CREATION UPDATE:**
![Screenshot 2024-03-07 at 3 18 27 PM](https://github.com/hyperdxio/hyperdx/assets/4743932/a948ef19-9978-4a6f-9beb-3ee49f660aac)