Change Saved Search Name (#358)
Currently, we are unable to change the name of a saved search. This PR primarily addresses this issue by adding a new option that appears when clicking the Menu icon in the top right corner. Upon selection, users will be able to enter a new name for the saved search. After pressing confirm, the change will be successfully applied.

## Demo

https://github.com/hyperdxio/hyperdx/assets/5051756/8ca461a1-53b1-41d9-b057-31c1fe5e1b00