fix(dev): otel collector should wait for clickhouse
The OTEL collector can start execution before the Clickhouse DB
is ready. This can cause the OTEL collector to crash, which aborts
the container and caused the compose command to fail. This commit
adds a health check to the Clickhouse container and then orders the
OTEL collector container to wait until the database is healthy.