fix: Add tooltip for password re-entry for testing connections (#631)
Adds a tooltip on hover next to the test connections buttom that informs users that they need to re-enter their password in order for test connection to work properly.

Ref: HDX-1377

![image](https://github.com/user-attachments/assets/9e2741fa-cc7b-4fab-86ff-fb119528f112)


Tooltip appears to the right of the button only in edit mode. Confirmed that it does not appear for new connections.