Fix search UI and add suggestions to queries that failed (#693)
Fixed the UI for DBSearchPage when one of the queries error. Only one error is displayed instead of two. Error shows the user's inputs and the exception, in hopes that the exception will help the user debug their query.

To further help the user debug their query, I added a new file called "useSqlSuggestion.tsx" that has parsers to search through the form's select, where (but not for Lucene), and orderBy clauses to detect errors and offer suggestions to the user. The user can then click a button that will fix the appropriate input. So far the only suggestion offered would be changing double quotes (invalid for ClickHouse queries) to single quotes. Example: SeverityText:"err" => SeverityText:'err'

Ref: HDX-1489
Ref: HDX-1486
Ref: HDX-1267