feat: Handle JSON-like structured logs better
Updates the OTEL pipeline to handle structured logs better. If the body
content is an OTEL map, it will merge the body map into the log attributes
map. If the body is a JSON object, it will parse the JSON string into an
OTEL map, then merge the fields into the log attributes map.

Replacing the Body field doesn't work since the Clickhouse exporter schema
defines Body as string, so any parsed out object ends up turning back into
a string. At least as log resources, it's a lighter weight means of grouping
and filtering in the UI.

Ref: HDX-1453