fix: remove noisy log message
Now that the app has some complex queries that leverage CTEs, metrics
for example, it's common for the logic in this optimization to throw
an exception. When that happens, the query rendering logic continues
without a problem but generates a noisy line in the console log. We can
just remove this log message to clean up the debugging experience.

Ref: HDX-1763