feat: Improve source editing UX (#805)
Background: the gear icon on the search page showed a modal for editing the source. Due to low friction and lack of context clues, it appeared as if those settings are ad-hoc for the current query.

The new UX is a dropdown with a link to the Teams page to make it clear that those are team-wide settings.

![Screenshot 2025-05-10 at 8 16 24 PM](https://github.com/user-attachments/assets/c2280bed-a29c-460f-8243-5785cda87c3e)