feat: Add download csv functionality to search tables (#939)
Adds a download icon that allows users to download a csv of results. Note: In v1, this was a gear icon that brought up a modal with a few different search options, including adding additional columns. Since that functionality doesn't exist in v2 yet, I thought it was best to just have a direct icon for now.

![image](https://github.com/user-attachments/assets/f6c71d0e-951e-4cc8-a2af-489c03a53598)


Fixes: HDX-1590