review: use `id` instead of `_id` for entities
The property `_id` is a mongo specific detail that has leaked into
the code and models. A more readable and generic property, `id`,
should be used instead.

This commit adds the `id` property to a number of the alert entities,
adjust the mongoose model to serialize the virtual `id` field from
the documents and then updates usage in the alert check to use the
new property.