fix: Prevent empty order by set in search page for certain sort/primary keys (#1124)
When using a trace table with the default otel schema of `ServiceName, SpanName, toDateTime(Timestamp)`, we currently return a blank order by which breaks sorting in search page.

Instead, if we fail to optimize the order by based on the primary key, we should just return a default value instead of an empty value.