fix: simplify alert provider load logic
Dynamic module resolution causes issues in builds that use tools like
esbuild; resulting paths are often not the same as they would in the source
tree. Instead we can simplify the loading logic to use a defined object map
of names to constructor functions. The mapping object should be small
enough that merge conflicts with forks should be easy to resolve.