fix: Fix pattern sample query for sources with multi-column timestamp expressions (#1304)
Fixes HDX-2621

# Summary

This PR fixes a query error when opening a sample log line in the patterns table from a source with multiple timestamp columns. To fix the issue, we simply use the first of the timestamp columns. This is consistent with several other places in the app where we use just the first timestamp column - multiple timestamp columns is not fully supported.

## Testing

To reproduce the issue, set the Timestamp Column in source settings for the logs source to `TimestampTime, toStartOfMinute(TimestampTime)`. Then attempt to open a pattern sample:

https://github.com/user-attachments/assets/2464f97e-1423-437c-88f0-b45486feffcc

With these changes, the issue is fixed:

https://github.com/user-attachments/assets/54d8f0f2-532c-4eb4-a676-ab6a606ecac5