feat: Add service map (beta) (#1319)
Closes HDX-2699

# Summary

This PR adds a Service Map feature to HyperDX, based on (sampled) trace data.

## Demo

https://github.com/user-attachments/assets/602e9b42-1586-4cb1-9c99-024c7ef9d2bb

## How the service map is constructed

The service map is created by querying client-server (or producer-consumer) relationships from a Trace source. Two spans have a client-server/producer-consumer relationship if (a) they have the same trace ID and (b) the server/consumer's parent span ID is equal to the client/producer's span ID. This is accomplished via a self-join on the Trace table (the query can be found in `useServiceMap.ts`.

To help keep this join performant, user's can set a sampling level as low as 1% and up to 100%. Lower sampling levels will result in fewer rows being joined, and thus a faster service map load. Sampling is done on `cityHash64(TraceId)` to ensure that either a trace is included in its entirety or not included at all.