fix: Fix session page source change on submit (#1380)
Closes HDX-2858

# Summary

This PR fixes a bug on the sessions page which caused any search submission to revert the source select back to its default. 

The cause of this bug was that the form submission was updating the wrong URL query parameter (source instead of sessionSource), and form submission would trigger the select to take whatever (original) value was set for sessionSource).