feat: Add search to trace waterfall (#1377)
Closes HDX-2753
Closes HDX-2755
Closes HDX-2756

# Summary

This PR adds a search function for filtering spans and logs in the trace waterfall. Because the waterfall can consist of both spans and logs, from two different sources, there is one input per source. If the correlated log source is not available (ensuring there are no logs in the waterfall) then there is only one input.

The input persists in the query parameters so that the filtered side panel can be shared, and the parameters are cleared when the side panel closes. 

Currently, only Lucene is supported for searching.

This PR also adds a couple of minor improvements to the waterfall

1. There is now a count of spans and errors
2. There is now a span status in the waterfall tooltip 

## Demo


https://github.com/user-attachments/assets/fb623875-5811-4f7f-9f40-c0b34de1c541