fix: Disable useSessionId query when traceId input is undefined (#1451)
Closes #1449
Closes HDX-2992

# Summary

This PR prevents HyperDX from issuing an invalid query (described in #1449) by disabling the query for sessions when a trace ID is not available. This often happened not just for events without a trace ID but also immediately after opening the side panel while a valid traceId was still loading.

Updating the types to reflect the possibility that trace ID is undefined also exposed the fact that we try to show the trace waterfall and row overview panels on the trace panel even when there is no trace id. This has been fixed.

## Before

https://github.com/user-attachments/assets/91994d82-4c1c-4538-bc4b-0ee31480200a

## After

No invalid query is issued:

https://github.com/user-attachments/assets/300580d3-970e-405f-868b-d0aec9b722e7

When there is no trace id, we don't attempt to render the waterfall:
<img width="1342" height="832" alt="Screenshot 2025-12-05 at 10 02 14 AM" src="https://github.com/user-attachments/assets/80b87d02-2a80-49e2-a0ee-4808d712de0b" />