fix: Ensure adequate SQL/Schema Preview modal height (#1450)
Closes HDX-2865
Closes HDX-2980
Closes #1443 

# Summary

This PR ensures that the schema preview / SQL query preview modal does not have a restrictive max height.

Previously, the CodeMirror component would _sometimes_ have a max height of 150px applied. It appears that this was coming from the global `cm-editor` and `cm-scroller` classes, which were conditionally updated with max height properties in `SQLInlineEditor.tsx` for multi-line inputs.

To fix this, the SQLInlineEditor multiline support now makes use of separate CSS selectors to only apply the max height to inputs with allowMultiline enabled.

## Demo

Modals are now >150px reliably, and multi-line editors are still capped to 150px max height

https://github.com/user-attachments/assets/e3d69c3d-3657-42ef-ba88-a961f0b23cb4