feat: Filter on isRootSpan column if present (#1461)
Closes HDX-3009

# Summary

This PR updates the Root Spans Only filter to support a materialized `isRootSpan` column, which is expected to be present on some trace table schemas. If that column is present, then the `Root Spans Only` filter will add a filter like 'isRootSpan IN (TRUE)` to the query, instead of the default `ParentSpanId IN ('')`. The UI has also been updated to support displaying and pinning boolean filter values.

**Note:** We still only query string type filter values, so we won't show isRootSpan unless Root Spans Only has been toggled.

<details>
<summary>I confirmed that if `isRootSpan` is in the ordering key, then this new condition will utilize the key to prune granules:</summary>

<img width="1678" height="1065" alt="Screenshot 2025-12-11 at 11 54 35 AM" src="https://github.com/user-attachments/assets/e22ae689-25a9-4d6b-b0f6-cc8f8396c35b" />
</details>

## Demo

<details>
<summary>For a source with the isRootSpan filter</summary>

https://github.com/user-attachments/assets/ccc7a890-b16e-4de6-bbb9-295fb10aa214
</details>

<details>
<summary>For a source without the isRootSpan filter (no change)</summary>

https://github.com/user-attachments/assets/33d4dd0a-136a-4284-812c-ddd12e67246e
</details>