introduce parallelized chunking (#1460)
Some queries benefit from being windowed, but do not require being run in series. The histogram on top of the search page is the perfect example! This PR enables that query to be run in parallel. It does essentially the same thing as the code that runs it in series, but wraps in a `Promise.all`.

View the video here and compare the speed of loading a histogram. Compare by testing play.hyperdx.io and then the published preview.


https://github.com/user-attachments/assets/d519f643-9d84-4ed0-a8b7-84570f80a58a


Closes HDX-3030