fix(otel-collector): fix log rotation script
There were two issues with the log rotation script:

1. Logs could be lost since copying and then truncating the file might
   not finish before logs arrive.

2. The otel collector application will keeps the file handle and offset
   cached. After truncating, it will write starting at the last offset
   leaving the unallocated garbage in the beginning of the file. This
   garbage uses space.

This commit moves the file instead of copying. That allows the collector
to continue writing to the rolled file until a SIGHUP is sent. This causes
a config refresh, which also opens a new log file. After, the rolled file
and the new log file have correct sizes.