fix: support otel collector logging with minimal storage
This commit sets up a FIFO named pipe with the same name/path that the otel
collector and supervisor are expecting. By starting to tail that pipe before
starting the collector, we can send log files to stdio without the memory
required by the `passthrough_logs` feature or and storage on the volume.